fix(syscall): Corrupted args when dispatching

+++ src/include/wey/syscall.h
@@ -2,7 +2,6 @@
 #define _SYSCALLS_H
 
 #include <def/compile.h>
-#include <stdint.h>
 
 #define SYSCALL_INTERRUPT_NUM 0x80
 
@@ -54,7 +53,7 @@
 #define SYSCALL_DEFINE5(name, ...) SYSCALL_DEFINEx(5, _##name, __VA_ARGS__)
 #define SYSCALL_DEFINE6(name, ...) SYSCALL_DEFINEx(6, _##name, __VA_ARGS__)
 
-typedef asmlinkage long (*sys_fn_t)(long long, long long, long long, long long, long long, long long);
+typedef asmlinkage long (*sys_fn_t)(long, long, long, long, long, long);
 
 void syscalls_init();
